outfits: introduce {weapon,engine} space reducers
there are instances where one might have unused weapon or engine space
which cannot or won't be used at a time.
so like there are expansions to increase the weapon and engine space
in favor of outfit space, this commit adds the opposite.
e.g. expansions to increase outfit space in favor of weapon or engine space.

Signed-off-by: Dagg Stompler <daggs@gmx.com>